backend/requirements.txt
fastapi 
uvicorn
python-dotenv
python-multipart 
psycopg2-binary 
passlib[bcrypt] 
python-jose[cryptography] 
fastapi-jwt-auth 
cloudinary 
vaderSentiment
supabase
supabase[async]